appbar - separation concerns
