+++ lib/page/mytests/appbar_responsiveness/homepage_appbar_responsiveness.dart
@@ -1,11 +1,13 @@
 import 'dart:async';
 import 'dart:math';
 
-import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
-import 'package:riverpod_modifiers_example/page/mytests/appbar_responsiveness/myapp_appbar_responseive.dart';
+import 'package:riverpod_modifiers_example/page/mytests/appbar_responsiveness/widgets/appbar_content.dart';
+import 'package:riverpod_modifiers_example/page/mytests/appbar_responsiveness/widgets/body_content.dart';
 
 import 'package:riverpod_modifiers_example/widget/button_widget.dart';
+
+import 'myapp_appbar_responseive.dart';
 
 class HomePageAppBarResponsiveness extends StatefulWidget {
   @override
@@ -13,8 +15,8 @@ class HomePageAppBarResponsiveness extends StatefulWidget {
 }
 
 class _HomePageAppBarResponsivenessState extends State<HomePageAppBarResponsiveness> {
-  double appBarHeight = 200;
-  String lixo = 'responsiveness';
+  String _buttonText = 'init...';
+  double _appBarHeight = 100;
 
   @override
   void initState() {
@@ -25,134 +27,30 @@ class _HomePageAppBarResponsivenessState extends State<HomePageAppBarResponsiven
   _changeAppbarHeight() {
     setState(() {
       final randomValue = Random().nextInt(100) + 100;
-      appBarHeight = randomValue.toDouble();
-      lixo = 'bosta $randomValue';
+      _appBarHeight = randomValue.toDouble();
+      _buttonText = 'random $randomValue';
     });
   }
 
   _startTimer() async {
     Timer.periodic(Duration(seconds: 2), (timer) {
       _changeAppbarHeight();
-      // myLogger.d('${timer}');
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    return _responsiveness();
-  }
-
-  Widget _responsiveness() {
     return Scaffold(
-      appBar: buildAppBar(context),
+      appBar: appBarContentFactory(_appBarHeight).appBar,
       body: Center(
         child: SingleChildScrollView(
           child: Column(
             children: [
-              ButtonWidget(text: lixo, onClicked: () {}),
-              buildSimpleTextcreen(),
+              bodyContent(_buttonText),
             ],
           ),
         ),
       ),
     );
   }
-
-  final mainItemSize = 40.0;
-  final subItemSize = 20.0;
-
-  SliverAppBar buildSilverAppBar(BuildContext context) {
-    return SliverAppBar(
-      floating: true,
-      leading: IconButton(
-        icon: Icon(
-          Icons.arrow_back_rounded,
-        ),
-        onPressed: () => Navigator.of(context).pop(),
-      ),
-      flexibleSpace: _column([
-        _mainItem(),
-        _subItem(),
-      ]),
-      // expandedHeight: 200,
-    );
-  }
-
-  AppBar buildAppBar(BuildContext context) {
-    return AppBar(
-      toolbarHeight: appBarHeight,
-      leading: IconButton(
-        icon: Icon(
-          Icons.arrow_back_rounded,
-        ),
-        onPressed: () => Navigator.of(context).pop(),
-      ),
-      title: _column([_mainItem(), _subItem()]),
-    );
-  }
-
-  Widget _expanded(Widget w) {
-    return Expanded(
-      child: w,
-    );
-  }
-
-  Widget _column(List<Widget> wList) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [...wList],
-    );
-  }
-
-  bool isNormal = true;
-
-  Widget _mainItem() {
-    return _getTextGlobal(_mainContent(), mainItemSize, isNormal);
-  }
-
-  Widget _subItem() {
-    return _getTextGlobal(_subContent(), subItemSize, isNormal);
-  }
-
-  Widget _getTextGlobal(String content, double size, bool isNormal) {
-    if (isNormal) {
-      return _getText(content, size);
-    } else {
-      return _getAutoSizeText(content, size);
-    }
-  }
-
-  Widget _getText(String content, double size) {
-    return Text(
-      content,
-      maxLines: 3,
-      overflow: TextOverflow.ellipsis,
-      style: TextStyle(fontSize: size),
-    );
-  }
-
-  Widget _getAutoSizeText(String content, double size) {
-    return AutoSizeText(
-      content,
-      maxLines: 3,
-      overflow: TextOverflow.ellipsis,
-      style: TextStyle(fontSize: size),
-    );
-  }
-
-  String _mainContent() {
-    return 'Main Lixooooooooo 11111111111111 rrrrrrrrrrrrr rrrrrrrrrrrrrr rrrrrrrrrrrrrr  pppppppppp ooooooooooooo';
-  }
-
-  String _subContent() {
-    return 'Sub teste 12345';
-  }
-
-  final buildSimpleTextcreen = () => Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          //  _SimpleText(content: 'golden'),
-        ],
-      );
 }
