+++ lib/page/mytests/appbar_responsiveness/widgets/body_content.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:riverpod_modifiers_example/widget/button_widget.dart';
+
+final bodyContent = (String buttonText) => Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        ButtonWidget(text: buttonText, onClicked: () {}),
+        //  _SimpleText(content: 'golden'),
+      ],
+    );
